Close #807. Consolidate/simplify component dependencies
