+++ R/card.R
@@ -72,7 +72,6 @@ card <- function(..., full_screen = FALSE, height = NULL, max_height = NULL, min
     !!!attribs,
     !!!children,
     if (full_screen) full_screen_toggle(),
-    card_dependency(),
     card_init_js()
   )
 
@@ -279,17 +278,6 @@ full_screen_toggle <- function() {
     ),
     "Expand"
   )
-}
-
-card_dependency <- function() {
-  list(
-    component_dependency_js("card"),
-    bs_dependency_defer(card_dependency_sass)
-  )
-}
-
-card_dependency_sass <- function(theme) {
-  component_dependency_sass(theme, "card")
 }
 
 card_init_js <- function() {
