+++ R/sidebar.R
@@ -250,7 +250,6 @@ layout_sidebar <- function(
       "--bslib-sidebar-max-height-mobile" = max_height_mobile
     ),
     !!!contents,
-    sidebar_dependency(),
     sidebar_init_js()
   )
 
@@ -311,16 +310,4 @@ sidebar_init_js <- function() {
     `data-bslib-sidebar-init` = NA,
     HTML("bslib.Sidebar.initCollapsibleAll()")
   )
-}
-
-
-sidebar_dependency <- function() {
-  list(
-    component_dependency_js("sidebar"),
-    bs_dependency_defer(sidebar_dependency_sass)
-  )
-}
-
-sidebar_dependency_sass <- function(theme) {
-  component_dependency_sass(theme, "sidebar")
 }
