+++ R/utils.R
@@ -36,7 +36,6 @@ path_inst <- function(...) {
 }
 
 path_lib <- function(...) path_inst("lib", ...)
-path_components <- function(...) path_inst("components", ...)
 
 is_shiny_app <- function() {
   # Make sure to not load shiny as a side-effect of calling this function.
