+++ R/value-box.R
@@ -157,8 +157,7 @@ value_box <- function(
       "--bslib-color-bg" = theme$bg
     ),
     !!!attribs,
-    contents,
-    as.card_item(value_box_dependency())
+    contents
   )
 
   as_fragment(tag_require(res, version = 5, caller = "value_box()"))
@@ -196,17 +195,8 @@ value_box_auto_border_class <- function(theme, class = NULL) {
   return(NULL)
 }
 
-value_box_dependency <- function() {
-  bs_dependency_defer(value_box_dependency_sass)
-}
 
-value_box_dependency_sass <- function(theme) {
-  deps <- component_dependency_sass(theme, "value_box")
-  # Inject the icon gradient SVG into the dependency $head slot
-  deps$head <- value_box_dependency_icon_gradient()
-  deps
-}
-
+# TODO: bring in gradient as a tags$head(singleton())? Or just with the bslib JS init?
 value_box_dependency_icon_gradient <- local({
   cached <- NULL
