+++ srcts/src/components/webcomponents/index.ts
@@ -1,8 +1,8 @@
 import { BslibTooltip } from "./tooltip";
 import { BslibPopover } from "./popover";
 import { BslibInputDarkMode } from "./inputDarkMode";
-import { makeInputBinding } from "./webcomponents/_makeInputBinding";
-import { shinyAddCustomMessageHandlers } from "./webcomponents/_shinyAddCustomMessageHandlers";
+import { makeInputBinding } from "./_makeInputBinding";
+import { shinyAddCustomMessageHandlers } from "../_shinyAddCustomMessageHandlers";
 
 [BslibTooltip, BslibPopover, BslibInputDarkMode].forEach((cls) => {
   customElements.define(cls.tagName, cls);
